requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2
PyJWT==2.8.0
werkzeug==3.0.1

